requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
joblib>=1.3.0
insightface>=0.7.3
scikit-learn>=1.3.0
onnxruntime>=1.15.0
